backend/ocr/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
pillow==11.0.0
pytesseract==0.3.13
